+++ src/enums/PayableEntity.ts
@@ -2,7 +2,8 @@ enum PayableEntity {
   ApplePay = 'applebtn',
   GooglePay = 'googlebtn',
   Form = 'form',
-  Resign = 'resign'
+  Resign = 'resign',
+  Paypal = 'paypal'
 }
 
 export default PayableEntity