+++ src/interfaces/messages/OrderStatusMessage.ts
@@ -2,8 +2,11 @@ import Message from './Message'
 import OrderStatus from '../OrderStatus'
 
 import MessageType from '../../enums/MessageType'
+import PayableEntity from '../../enums/PayableEntity'
+import APMOrderStatus from '../APMOrderStatus'
 
 export default interface OrderStatusMessage
   extends Message<MessageType.OrderStatus> {
-  response: OrderStatus
+  entity: PayableEntity
+  response: OrderStatus | APMOrderStatus
 }